relocate the model upload button for human detection
