+++ exercises/templates/exercises/human_detection/exercise.html
@@ -44,7 +44,6 @@
                         {% end_react_component%}
                         {% react_component components/wrappers/FlexContainer%} 
                         {% react_component exercise/human_detection/SpecificBasicComputerVision %}{% end_react_component %}
-                        {% react_component components/visualizers/FileUploadComponent %}{% end_react_component %}
                             {% react_component components/visualizers/Camera%}{% end_react_component %}
                         {% end_react_component %}
                 {% end_react_component %}
